+++ src/app/(dashboard)/instagram-feed/components/form/media/load-button.tsx
@@ -1,0 +1,34 @@
+import type { FetchNextPageOptions, InfiniteQueryObserverResult, InfiniteData } from '@tanstack/react-query';
+
+import { Button } from '@/components/ui/button';
+import { Spinner } from '@/components/ui/spinner';
+
+import type { InstagramFeedFormMedia } from '../../../schemas/form.schema';
+
+export default function InstagramFeedFormMediaListLoadButton({
+  fetchNextPage,
+  isFetchingNextPage,
+}: {
+  fetchNextPage: (
+    options?: FetchNextPageOptions | undefined,
+  ) => Promise<InfiniteQueryObserverResult<InfiniteData<InstagramFeedFormMedia | undefined, unknown>, Error>>;
+  isFetchingNextPage: boolean;
+}) {
+  return (
+    <Button
+      type="button"
+      onClick={() => fetchNextPage()}
+      disabled={isFetchingNextPage}
+      className="flex cursor-pointer self-center"
+    >
+      {isFetchingNextPage ? (
+        <span className="flex items-center gap-1">
+          <Spinner data-icon="inline-start" />
+          Loading...
+        </span>
+      ) : (
+        'Load more'
+      )}
+    </Button>
+  );
+}