+++ src/app/(dashboard)/instagram-feed/components/list/table.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { DataTable } from '@/components/ui/data-table';
-import type { InstagramFeed } from '@/schemas/instagram-feed.schema';
 
+import type { InstagramFeed } from '../../schemas/feed.schema';
 import InstagramFeedListColumns from './columns';
 
 export default function InstagramFeedListTable({ feedList }: { feedList: InstagramFeed[] }) {