+++ src/app/(dashboard)/instagram-feed/hooks/use-create-form.ts
@@ -5,12 +5,11 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
-import { useDashboardUser } from '@/app/(dashboard)/user-provider';
-import { InstagramFeedFormSchema } from '@/schemas/instagram-feed.schema';
-import type { InstagramFeedForm } from '@/schemas/instagram-feed.schema';
-import { useInstagramFeedStore } from '@/store/instagram-feed-store';
-
-import { createInstagramFeed } from '../../actions/create-feed.action';
+import { useDashboardUser } from '../../user-provider';
+import { useInstagramFeedStore } from '../store/instagram-feed-store';
+import { createInstagramFeed } from '../actions/create-feed.action';
+import { InstagramFeedFormSchema } from '../schemas/form.schema';
+import type { InstagramFeedForm } from '../schemas/form.schema';
 
 export default function useInstagramFeedCreateForm() {
   const { user } = useDashboardUser();